order scripts, better error checking

+++ scripts/startup.sh
@@ -1,24 +1,22 @@
 #!/bin/sh
 
-# TODO check installed
-# if [ -n "$(cat /etc/hosts | grep vhcalnplci.dummy.nodomain )" ]; then
-#   echo "Seems the system configs already patched, skipping"
-#   exit 0
-# fi
-
-echo "Local start/stopsap scripts..."
 _LOCAL_BIN=/usr/local/bin
-cp /vagrant/scripts/startsap.sh $_LOCAL_BIN
-cp /vagrant/scripts/stopsap.sh $_LOCAL_BIN
-chmod 755 $_LOCAL_BIN/startsap.sh
-chmod 755 $_LOCAL_BIN/stopsap.sh
-
-echo "Installing service..."
 _SERVICE_PATH=/etc/systemd/system/sapnw.service
-cp /vagrant/scripts/sapnw.service /etc/systemd/system/
-chmod 644 $_SERVICE_PATH
-systemctl daemon-reload
-systemctl enable sapnw
-systemctl status sapnw --no-pager --full
 
-# systemctl start sapnw
+if [ ! -e $_LOCAL_BIN/startsap.sh ]; then
+    echo "Local start/stopsap scripts..."
+    cp /vagrant/scripts/bin/startsap.sh $_LOCAL_BIN
+    cp /vagrant/scripts/bin/stopsap.sh $_LOCAL_BIN
+    chmod 755 $_LOCAL_BIN/startsap.sh
+    chmod 755 $_LOCAL_BIN/stopsap.sh
+fi
+
+if [ -e /usr/sap/NPL/D00/exe/sapstart ] && [ ! -e $_SERVICE_PATH ]; then
+    echo "Installing service..."
+    cp /vagrant/scripts/sapnw.service /etc/systemd/system/
+    chmod 644 $_SERVICE_PATH
+    systemctl daemon-reload
+    systemctl enable sapnw
+    systemctl status sapnw --no-pager --full
+    # systemctl start sapnw
+fi
